backend/requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
sentinelhub>=3.10.0
python-dotenv>=1.0.0
requests>=2.31.0
numpy>=1.26.0  # Must be >= 1.26 for Python 3.13
google-generativeai>=0.3.1
Pillow>=10.1.0
